MyOOSDumper 5.0.2 Revision of the directory protection function
Note: The changes listed below have increased the minimum PHP version from 5.3 to 5.5.

* To check whether a directory protection exists, the existence of `.htaccess` is no longer checked, but the effective directory protection is tested with a new function. In doing so, an external access to MOD is simulated.
  * New function `IsAccessProtected()` added.
* The control for the display of whether a directory protection exists and the control for the generation of the buttons for the directory protection are now independent mechanisms.
* MOD now not only knows the status "Directory protection not available", but also the status "Directory protection available", "Directory protection incomplete" and "Higher-level permissions detected".
  * Accordingly, 3 new language variables have been added.
* Directory protection generator revised and adapted to currently valid methods.
  * Removed Methods: MD5
  * Added methods: MD5APR (replaces MD5), bcrypt
  * Accordingly, 1 new language variable has been added.
  * Method descriptions changed.
  * List of methods rearranged.
  * Implemented the MD5APR algorithm from WhiteHat101.
  * The standard hash method for Linux-based systems is no longer CRYPT, but MD5APR, which has also been the standard for Apache since 2.2.18.
  * The generated content of `.htaccess` and `.htpasswd` is now contained in `<pre> ... </pre>` tags. This means that no HTML line breaks have to be inserted in the displayed content. Furthermore, the version switch for PHP 5.3 and XHTML has been removed.
  * When deleting the directory protection you will now be asked.
  * Accordingly, 1 new language variable has been added.

Secondary changes:
* The designation of the start page (heading) is no longer permanently entered in the source. Instead, the existing language variable `L_HOME` is now used.
* The language variable `L_HTACC_CONFIRM_DELETE` was named incorrectly and was renamed to `L_HTACC_CONFIRM_CREATE`.
* Changes and corrections to several language files.

+++ msd/tpl/home/home.tpl
@@ -3,13 +3,30 @@
 	{DIRECTORY_WARNINGS.MSD}
 <!-- END DIRECTORY_WARNINGS -->
 
+<!-- BEGIN HTACCESS_NOT_NEEDED -->
+	{L_HTACC_NOT_NEEDED}
+<!-- END HTACCESS_NOT_NEEDED -->
+
+<!-- BEGIN HTACCESS_COMPLETE -->
+	{L_HTACC_COMPLETE}
+<!-- END HTACCESS_NOT_NEEDED -->
+
+<!-- BEGIN HTACCESS_INCOMPLETE -->
+	<span class="error">{L_HTACC_INCOMPLETE}</span>
+<!-- END HTACCESS_INCOMPLETE -->
+
+<!-- BEGIN HTACCESS_NEEDED -->
+	<span class="error">{L_HTACC_PROPOSED}</span>
+<!-- END HTACCESS_NEEDED -->
+
 <!-- BEGIN HTACCESS_EXISTS -->
+	<br><br>
 	<a href="main.php?action=edithtaccess" class="Formbutton">{L_HTACC_EDIT}</a>&nbsp;
-	<a href="main.php?action=deletehtaccess" class="Formbutton">{L_DELETE_HTACCESS}</a><br>
+	<a href="main.php?action=deletehtaccess" class="Formbutton" onclick="if (!confirm('{L_HTACC_CONFIRM_DELETE}')) return false;">{L_DELETE_HTACCESS}</a><br>
 <!-- END HTACCESS_EXISTS -->
 
 <!-- BEGIN HTACCESS_DOESNT_EXISTS -->
-	<span class="error">{L_HTACC_PROPOSED}:</span>&nbsp;&nbsp;
+	<br><br>
 	<a href="main.php?action=schutz" class="Formbutton">{L_HTACC_CREATE}</a><br>
 <!-- END HTACCESS_DOESNT_EXISTS -->
